Improve parsing of namespace-prefixed symbols
Generate namespace tags for namespace-prefixed tags, such as
class Foo.Bar.Baz {}.

The support is still hackish and limited since it reads the namespace
prefixes as part of the tag's name and later split that tag name
accordingly. The main drawback of using this method is that the
namespace prefix cannot be separated from the tag name by blanks as any
other token can.

However, parsing these naturally seems a bit hard to do because the C
parser expects sub-blocks to be surrounded by braces.

Also allow blanks between the tilde and the name of the class in
destructors.