Fix win32_expand_environment_variables() and calling code encoding usage
win32_expand_environment_variables() use not to be Unicode-safe, but could
called with UTF-8 input.  Fix it to be Unicode-safe and to take input and
output in UTF-8.

Also, update calling code to always pass in UTF-8 data.  This required a
partial spawn code refactoring, which should now handle non-ASCII
characters better.