Only automatically indent upon newline if the line has default indent
Only automatically trigger indentation of the line upon newline if the
line has the same indent as it would have had when created (e.g. the
indent after line -1), so it is unlikely to modify an user indent but
still applies indentation.

If the line indentation should be re-computed upon newline no matter
its current indentation, one should use \r and \n triggers.