Use the GTK AccelMap to manage menu items accelerator labels
Use the proper GTK API for managing menu items accelerators, which
saves us from having to do it.  Also, the GTK code can seamlessly
manage any number of items sharing the same keybinding, where our code
used to only be able to really manage one, leading to the editor popup
menu not to be updated after startup.

So, this also fixes the update of the editor popup's keybindings, that
now properly displays the current bound accelerator.

FIXME: GtkAccelMap doesn't support dupliactes, but we allow them, so what?
Is the dupliacte keybinding a real feature?  Is that useful?