Speed up tag lookup in goto_tag()
Since we know tm_tags_extract() performs a linear search anyway, don't
use it to filter the array to search linearly in.  Instead, search
linearly directly in the whole array and stop on first match.

This avoids walking the whole array once even if the matching tag is
first, and also avoids creating a new temporary array for nothing.

This results in a 1.6x/2x speedup.