Speed up tag lookup in goto_tag()
Use tm_tags_find() and let it use binary search.  Use it also to find
workspace tags from the merged workspace tags instead of trying each
source file, again with binary search.

This results in speedups from 1.3x (when searching in current source
file and/or with the match close to the start) to more than 2000x (when
searching in all source files, with the match close to the end).