Let plugins fill GeanyPlugin::hooks instead of passing their own pointer
This is easier to handle if we decide to add hooks. Since we can zero-initialize
hooks before passing it to the plugin we can be certain as to what hooks
the plugin knew about when it was compiled. This is exactly the same method
used for GeanyPlugin::info already and easier than inspecting the API version.