Fix indentation after an alignment of the same width as the indentation
If the indentation after e.g. an open brace matches the previous line's
indentation with, but the previous indentation was an alignment, the
line should be indented as if it was a new level, not copying the exact
previous line's indentation.

This fixes e.g. an if's body indentation when indenting with tabs,
using a width of 4, using space alignment and having enabled
parentheses indentation:

	1.	if (foo &&
	2.	    bar) {
	3.		// ...
	4.	}

Line 3 should be indented with 2 tabs, but as the previous line's
indentation has the exact same effective width (with tabs of 4, 1 tab
plus 4 spaces) the logic trying to keep user's alignment would kick in
and lead to indenting line 3 with 1 tab and 4 spaces.

Fix this by only copying the exact previous line indentation if we
didn't mean to change the line's indentation.