parentheses-indent: Fix copy existing alignment when aligning
Properly copy the alignment of line L-1 when aligning line L, not to
incorrectly replace existing alignment with indentation in case of
nested alignment.

This only fixes the problem when aligning up (to larger width) but not
down (when aligning back to a matching parenthesis).

Of course, the problem is only visible when using non-space indentation
style, as otherwise alignment and indent use the same characters and so
can be mixed up without consequences.