parentheses-indent: Fix copying existing alignment when aligning back
Copy the indentation from the line on which the matching parenthesis
was found, and add appropriate alignment.

This fixes keeping existing alignment when aligning back, so nested
alignment works correctly.

FIXME: the implementation here is relatively ugly, as it doesn't
actually compute the indent/alignment, but relies on
insert_indent_after_line() to copy the existing line's exact indent.

While it is probably great to copy the exact indentation, the code
should probably either clearly rely on this, or yet try its best to
give proper indent and alignment, and so the "copy exact indent"
feature is only needed to keep user's custom indentation, not one the
code here would have created itself.

I don't know how important it is, but cleanup would be needed.