parentheses-indent: Don't backtrack paste a balanced line
This simplifies the logic a bit, by using the indentation of the first
balanced line, instead of searching for the real enclosing parenthesis
and then mitigating for manual user unindents.

This changes the behavior, as it will now use the indentation of the
first balanced line, rather than the smallest indent in the enclosing
parenthesis.  For example, it will give:

	1.	foo(1,
	2.	  bar(2, // manual unindent
	3.	      3),
	4.	    baz(3, // manual reindent
	5.	        4),
	6.	    5); // aligned to L4, closest balanced line

instead of the previous:

	1.	foo(1,
	2.	  bar(2, // manual unindent
	3.	      3),
	4.	    baz(3, // manual reindent
	5.	        4),
	6.	  5); // aligned to L2, smallest indent since L1