parentheses-indent: Only align up if no other indentation rule matched
Only apply parentheses indent if there is no other changes, or if it
would reduce alignment -- e.g. in case of

	1.	if (foo &&
	2.	    bar) {
	3.	  // here, needs to still reduce alignment
	4.	}

Not touching alignment if there are other changes allows not to align
in cases like e.g. JavaScript object literals:

	1.	foo({
	2.	  // no alignment
	3.	});

This looks sound as in such a case parentheses alignment would probably
be incorrect anyway, as it would align on the parenthesis and not the
open curly brace:

	1.	foo({
	2.	    // probably incorrect
	3.	});

It would however be interesting to be able to also align on curly
braces (and possibly square braces too) so we could handle cases like
that:

	1.	foo({ 1, 2, 3,
	2.	      4, 5, 6 });

Potentially, we might want to align on any kind of braces.  Not sure if
it's in the scope of this though.