travis: Split MinGW out of the default build matrix
We only want 1 job per environment and per compiler, so using the
environment to select the compiler is flawed as it will multiply when
adding compilers.

So, move this out of the default matrix, even though it means
duplicating the environment values.  But this isn't so bad, especially
as cross-compilation might only support a subset of what we could want
to select through the environment.