Don't show the goto popup for typedef synonyms
For instance when performing goto tag for Foo and Foo is defined as

typedef struct Foo {} Foo;

go immediately to the struct location without showing the goto popup with
both the struct name and typedef. When there are more occurrences of the
name, filter the list and don't show the synonyms in the popup.

In addition, if the cursor is on the same line as the typedef, go to
the struct and vice versa.

Note the missing

g_strcmp0(second->var_type, first->name) == 0

in the check - in this particular case we won't get the type to which the
typedef refers inside var_type because at the time the typedef tag is
generated in c.c the struct tag doesn't exist yet. On the other hand
there's no second->var_type == NULL either because this behaviour seems
to be rather implementation-specific and might easily change in the
future. The existing checks are probably sufficient for the real-world
code.