Filter symbols more thoroughtly in goto_tag()
Exclude types matching the current tag early instead of recursing if we
only found one candidate and it's on the current line, so we can do the
declaratiuon/definition swapping even with several matches.

This also fixes ignoring tags of the same name in the same file but
with different types.