scintilla: Fix fetching the frame border width on GTK 3.20
GTK 3.20 introduced a sub-node named "border" to the frame, requiring
to lookup the border on that sub-node rather than on the frame itself.

Unfortunately supporting this requires to be built against GTK 3.20, as
it requires API this version introduced.  This means that a build
against an older version won't be able to get the popup sizing right if
running on 3.20.