Fix typename highlighting before the visible area
We currently only re-highlights the visible area after typenames
changed.  This works well for the most common cases as typenames are
generally only used after their definition, but fails if a typename
change does affect some part of the document before the visible one
(i.e. in case of search & replace, or forward declarations).

Fix this by lazily re-highlighting the possibly outdated visible area
until we updated to the start of the document.