Hack to attempt to not use another user's configuration directory
This tries to determine whether the default configuration directory is
owned by the calling user, and falls back to a directory under that
user's home directory if not.

This doesn't solve the case where the first run would create the
directory in another user's home, but on the other hand handles any
combination of actual user and environment, not only root.

This however also doesn't work if the user is expected not to actually
own the files in hist home directory but e.g. only is part of the group
and still has write privileges.

Closes #294.
Closes #1150.
Closes https://sourceforge.net/p/geany/bugs/471/.