Make accessibility work on GTK < 3.8
On GTK < 3.2, we need to use the AtkObjectFactory.  We need to query
the factory to see what type we should derive from, thus making use of
dynamic inheritance.  It's tricky, but it works so long as it's done
carefully enough.

On GTK 3.2 through 3.6, we need to hack around because GTK stopped
registering its accessible types in the factory, so we can't query
them that way.  Unfortunately, the accessible types aren't exposed
yet (not until 3.8), so there's no proper way to know which type to
inherit from.  To work around this, we instantiate the parent's
AtkObject temporarily, and use it's type.  It means creating an extra
throwaway object and being able to pass the type information up to the
type registration code, but it's the only solution I could find.

On GTK 3.8 onward, we use the proper exposed GtkContainerAccessible as
parent, and so a straightforward class.

To hide and contain the complexity in type creation arising from the
hackish support for GTK 3.2 to 3.8, the actual implementation for the
widget's get_accessible() is located in the accessibility layer itself.

Inspiration for the GTK < 3.2 part comes from Evince 2.24, thanks.