Avoid restoring dialogs to positions outside existing monitors
Try and fit the dialogs inside an existing monitor when restoring their
placement in case the window manager allows placing windows outside any
monitor.

This is a typical problem on Windows where dialogs are restored to
offscreen locations after the monitor on which they were has been
disconnected, leading to having a hidden dialog.

Closes #1872 and https://sourceforge.net/p/geany/bugs/673/.